added iceberg tests
